Beter handle websockets when server is disconnecting due to network interface changes

+++ src/stores/bookmark.js
@@ -10,6 +10,10 @@ class Bookmark extends Store {
 		this.setState(initialState, 'socket_connect');
 
 		this.socket.on('disconnect', () => {
+			if (this.getStateProperty('configuringNetworkInterface')) {
+				return;
+			}
+			
 			if (this.getStateProperty('reboot') || this.getStateProperty('shutdown') || !_.isNull(this.getStateProperty('upgrade'))) {
 				return;
 			}
