+++ src/stores/user.js
@@ -12,6 +12,10 @@ class User extends Store {
 		this.setState(initialState, 'socket_connect');
 
 		this.socket.on('disconnect', () => {
+			if (this.getStateProperty('configuringNetworkInterface')) {
+				return;
+			}
+			
 			if (this.getStateProperty('reboot') || this.getStateProperty('shutdown') || !_.isNull(this.getStateProperty('upgrade'))) {
 				return;
 			}
